refactor(literals): make escaping a single implementation
escapeAppleScriptString/toAppleScriptLiteral duplicated AppleScriptBuilder's
private escapeString/formatValue/makeRecord line for line — two independent
answers to "format a JS value as AppleScript source" that could drift.

They cannot simply share the helpers module: helpers.ts imports builder.js for
tell(), so having builder.ts import back would rebuild the cycle removed in
#87. Put the primitives in literals.ts, which imports only types, and have the
builder, helpers, and sources/* all import from there. helpers.ts re-exports
them, so the published surface is unchanged.

Also route activate/isRunning through runScriptOrThrow rather than repeating
its throw logic, and document that both throw ScriptExecutionError.

Raised in review of #88.
